Environments/requirements_onnx.txt
# requirements_onnx.txt
# Core for ONNX-based projects (inference, analysis)
# Install: 
#   1. conda create -n onnx python=3.10; conda activate onnx; 
#   2. conda install cudatoolkit=11.8 cudnn=8.9 -c conda-forge; 
#   3. pip install -r requirements_onnx.txt -i https://pypi.tuna.tsinghua.edu.cn/simple

# 0.
onnx>=1.14.0,<1.17.0
onnxruntime-gpu>=1.16.0,<1.19.0
onnxsim>=0.4.33,<0.5.0
tqdm>=4.60.0,<5.0.0
#dllist>=2.0.0

# 1. Visualization
matplotlib>=3.5.0,<3.9.0
netron>=7.0.0

# 2. Scientific base
numpy>=1.22.0,<1.26.0
scipy>=1.8.0,<1.12.0
#pandas>=1.5.0,<2.2.0

# 3. FLOPs / MACs analysis
onnx-tool>=0.7.0,<1.0.0



# requirements_onnx_vision.txt
# Vision-specific extras (image processing, metrics, etc.)
# pip install -r requirements_onnx_vision.txt  (after requirements_onnx.txt)

# 1. Image Processing
opencv-python>=4.5.0,<5.0.0
Pillow>=8.4.0,<11.0.0
scikit-image>=0.19.0,<0.21.0
